requirements.txt
Flask==2.3.2
Flask_Login==0.6.2
itsdangerous==2.1.2
numpy>=1.24.3
pymongo==4.3.3
python_bcrypt==0.3.2
PyYAML==6.0
redis==4.5.5
sphinx_rtd_theme==1.2.1
gunicorn>=20.1.0
